requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
frozenlist==1.7.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
pillow==11.3.0
priority==2.0.0
propcache==0.3.2
Quart==0.20.0
requests==2.32.4
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.1
